requirements.txt
django==5.2
django-filter
django-widget-tweaks
django-compressor
django-crispy-forms
crispy-tailwind
django-cotton
